Implement square Hebb equation and set a 2 neurons set-up
The electric pulses are now sent from the soma, as it is the goal
where the growth cone should be directed.

Also, there is a wait time for initializing the parent-child
relationship between soma and growth cone agents, in order to avoid
a bug in RANA execution.

The electric pulses and soma triggering events have now different
names, as they represent different situations