Apply Hebb's learning rule to the growth intensity
The rule is applied every time a growth come receives an incoming
electric pulse. The typical time is established as 20ms. Hence, the
pulses delays are lowered from 100 to 20 ms, as it is a more
reasonable time, based on neuroscience papers.

The axon length is lowered down in the growth cone, so more axon
agents will spawn in the map.