Log cone-soma connections to a .csv file in a nested folder
A table in master agent is created, that registers the connections
between cones and somas by using Lua events.

This is done by using the Lua io.open call. When the master agent
is cleaned up, the table containing the connections between parents
and children is writen into the file.